Move some util functions to utils/
