+++ packages/core/src/sessions/http/utils/check-user-id-type.ts
